Overhaul the configure machinery and use the Autoconf macros developed for other packages (with modifications to let Kerberos support be optional).
Build with Kerberos support by default if Kerberos libraries are
found.  To disable Kerberos support, pass --without-krb5 to
configure.  The option to specify the Kerberos library location is now
--with-krb5, not --with-kerberos.

Add support for AIX's bundled Kerberos.  Thanks to Markus Moeller for
the porting information.

Define _ALL_SOURCE on AIX to get a prototype for vsnprintf.