Switch to Automake and Libtool
pam-afs-session is now built using Automake and Libtool to bring it
more in line with other software packages.  This means that it now
relies on Libtool to know how to generate a loadable module rather
than hand-configured linker rules.  This may improve portability on
some platforms and may hurt it on other platforms.