Separate setting defaults from either argument parsing function
Both the krb5.conf parsing function and the argument parsing function
were setting all the defaults, which meant that the second one called
would override all the results of the first.  Separate out setting
defaults to a different function that has to be called before either
and expand the test suite to explicitly test parsing an empty set of
arguments after the defaults are set.