Build the fake PAM library against the system headers
Don't try to replace the PAM system headers.  Instead, we'll try
to make all of our functions match the system header prototypes,
which may be painful.  We can get Linux working at least.  This
removes the need to build various versions of the code with the
TESTING define and allows us to test exactly the code that's in the
final module.

In the process, pass some more arguments into pam_start where
required by the attributes on the function definitions on Linux.