Clean up linker flags and install location for libtool
On Linux, if configured with a prefix of /usr (rather than /usr/local,
the default), the module will be installed into /lib/security (or
/lib32/security or /lib64/security if they exist) rather than
/usr/lib/security to match the default PAM configuration.

The module is now installed under $libdir/security except for the
above exception, rather than a lib32 or lib64 directory, except for
the special case of /usr.  To install into another lib32 or lib64
directory, use the --libdir option to configure.

Reduce the various compiler and linker flags to only those bits that
seem to still be relevant now that we're using libtool.