Recreate the PAG and tokens if they were lost
If there is no PAG, create a new one and obtain tokens in pam_setcred
and pam_open_session even if the module has already run.  This works
around destruction of the PAG on Linux by keyring initialization
modules, which can otherwise be hard to avoid due to the ordering
between the auth and session stacks.  This support uses the
VIOC_GETPAG AFS system call if it is available and falls back on
analyzing the supplemental group list if it is not.