Fall back on KRB5CCNAME in normal environment if not set in PAM
If there is no KRB5CCNAME environment variable set in the PAM
environment but there is one set in the regular environment, still run
aklog and use the KRB5CCNAME from the general environment.  This
provides correct behavior in cases such as password-less sudo with the
KRB5CCNAME environment variable preserved.