Rewrite the module/basic test to use PAM scripts
Import the rest of PAM scripted testing from rra-c-util, add some
additional infrastructure required to use it for this module, and
convert the basic test to use the new mechanism.  This still uses
the hack to simplify input logging.  I'll back out of that as soon
as the entire test suite has been converted.