Remove output testing from module/pag
This reduces our test coverage somewhat, although the behavioral
testing is still in place.  It's too hard to test output without
the script framework, but it currently doesn't support a callback
after every PAM step.  That won't be hard to add.  If I do add that,
the output testing can be added back in.